refactor: modify directory structure
Group coding challenges into directories for efficiency
