+++ src/com/hackerrank/datastructures/easy/HR1ReverseArray.java
@@ -1,4 +1,4 @@
-package com.hackerrank.easy.datastructures;
+package com.hackerrank.datastructures.easy;
 
 import java.util.Arrays;
